requirements.txt
aniso8601==7.0.0
appdirs==1.4.3
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-cors-headers==3.1.1
django-extensions==2.2.5
django-filter==2.2.0
django-graphql-jwt==0.2.2
django-guardian==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
graphene==2.1.8
graphene-django==2.6.0
graphene-file-upload==1.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
idna==2.8
Markdown==3.1.1
mysql-connector==2.2.9
mysqlclient==1.4.4
nltk==3.4.5
numpy==1.17.1
oauthlib==3.1.0
Pillow==6.2.1
promise==2.2.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
robotframework==3.1.2
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==4.1.0
Rx==1.6.1
selenium==3.141.0
singledispatch==3.4.0.3
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
tqdm==4.37.0
urllib3==1.25.6
webdrivermanager==0.7.4
